+++ unlock-js/src/__tests__/v10/updateKeyPrice.test.js
@@ -1,5 +1,5 @@
 import { ethers } from 'ethers'
-import * as UnlockV10 from 'unlock-abi-1-0'
+import * as UnlockV10 from '@unlock-protocol/unlock-abi-3'
 import utils from '../../utils'
 import { ZERO } from '../../constants'
 import NockHelper from '../helpers/nockHelper'
