+++ unlock-js/src/__tests__/v11/withdrawFromLock.test.js
@@ -1,5 +1,5 @@
 import { ethers } from 'ethers'
-import * as UnlockV11 from 'unlock-abi-1-1'
+import * as UnlockV11 from '@unlock-protocol/unlock-abi-4'
 import utils from '../../utils'
 import TransactionTypes from '../../transactionTypes'
 import NockHelper from '../helpers/nockHelper'
