+++ unlock-js/src/__tests__/v12/getLock.test.js
@@ -1,4 +1,4 @@
-import * as UnlockV12 from 'unlock-abi-1-2'
+import * as UnlockV12 from '@unlock-protocol/unlock-abi-5'
 import { ethers } from 'ethers'
 import Web3Service from '../../web3Service'
 import NockHelper from '../helpers/nockHelper'
@@ -60,7 +60,7 @@ describe('v12', () => {
         resultEncoder.encode(['uint256'], [3])
       )
 
-      // retrieve the bytecode and compare to v01
+      // retrieve the bytecode and compare to v12
       nock.ethGetCodeAndYield(
         lockAddress,
         UnlockV12.PublicLock.deployedBytecode
@@ -69,7 +69,7 @@ describe('v12', () => {
       // get the block number
       nock.ethBlockNumber(1337)
 
-      // retrieve the bytecode and compare to v01
+      // retrieve the bytecode and compare to v12
       nock.ethGetCodeAndYield(
         lockAddress,
         UnlockV12.PublicLock.deployedBytecode
