+++ unlock-js/src/__tests__/walletService.test.js
@@ -3,7 +3,7 @@ import http from 'http'
 import NockHelper from './helpers/nockHelper'
 
 import v0 from '../v0'
-import v01 from '../v01'
+import v1 from '../v1'
 import v02 from '../v02'
 import v10 from '../v10'
 import v11 from '../v11'
@@ -15,7 +15,7 @@ import utils from '../utils'
 import WalletService from '../walletService'
 import { GAS_AMOUNTS } from '../constants'
 
-const supportedVersions = [v0, v01, v02, v10, v11, v12, v13, v7]
+const supportedVersions = [v0, v1, v02, v10, v11, v12, v13, v7]
 
 const endpoint = 'http://127.0.0.1:8545'
 const nock = new NockHelper(endpoint, false /** debug */)
