+++ unlock-js/src/v1/isLockManager.js
